update resource_monitor.bindings.py to py3
