+++ resource_monitor/src/bindings/python2/example_simple_limit.py
@@ -1,0 +1,1 @@
+../python3/example_simple_limit.py
